+++ CHANGELOG.md
@@ -5,6 +5,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Dates in this file meets Gregorian calendar. Date in format YYYY-MM-DD.
 
+## [0.0.3] - [2025-11-NN], Developer Release
+
+### Added
+
+- Selfie window.
+- Localization switcher.
+
+### Removed
+
+- Main menu.
+
+### Updated
+
+- Localization.
+- CPL dependency to v1.6.1.
+
 ## [0.0.2] - [2025-11-19], Developer Release
 
 ### Added